Merged from Master

+++ src/pm/pm_mpi_fft.h
@@ -12,6 +12,8 @@
 #ifndef PM_MPI_FFT_H
 #define PM_MPI_FFT_H
 
+#include "gadgetconfig.h"
+
 #include "../mpi_utils/setcomm.h"
 
 #ifndef FFTW
@@ -63,11 +65,7 @@ class pm_mpi_fft : public virtual setcomm
 
     int transposed_firstcol, transposed_ncol;
     int second_transposed_firstcol, second_transposed_ncol;
-    size_t second_transposed_ncells;
-
-    //  int pivotcol; /* to go from column number to task */
-    //   int avg;
-    //   int tasklastsection;
+    long long second_transposed_ncells;
 
     size_t *offsets_send_A;
     size_t *offsets_recv_A;
